+++ lib/lightstreamer/cli/commands/stream_command.rb
@@ -12,6 +12,7 @@ module Lightstreamer
       option :items, type: :array, required: true, desc: 'The names of the item(s) to stream'
       option :fields, type: :array, required: true, desc: 'The field(s) to stream'
       option :mode, enum: %w(distinct merge), default: :merge, desc: 'The operation mode'
+      option :selector, desc: 'The selector for table items'
       option :maximum_update_frequency, type: :numeric, desc: 'The maximum number of updates per second for each item'
 
       def stream
@@ -37,13 +38,18 @@ module Lightstreamer
 
       # Creates a new subscription from the specified options.
       def create_subscription
-        subscription = Lightstreamer::Subscription.new items: options[:items], fields: options[:fields],
-                                                       mode: options[:mode], adapter: options[:adapter],
-                                                       maximum_update_frequency: options[:maximum_update_frequency]
+        subscription = Lightstreamer::Subscription.new subscription_options
 
         subscription.add_data_callback(&method(:subscription_data_callback))
 
         subscription
+      end
+
+      def subscription_options
+        {
+          items: options[:items], fields: options[:fields], mode: options[:mode], adapter: options[:adapter],
+          maximum_update_frequency: options[:maximum_update_frequency], selector: options[:selector]
+        }
       end
 
       def subscription_data_callback(_subscription, item_name, _item_data, new_values)