+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.4 — Unreleased
 
-- Added support for specifying a subscription's maximum update frequency
+- Added support for specifying a subscription's selector and maximum update frequency
 - Added `Lightstreamer::Session#force_rebind` which asks the server to send a `LOOP` message so the client has to rebind 
   using a new stream connection
 - Added validation of the arguments for control connection subscription requests